首页热力图接口修改

+++ src/main/java/com/hydata/intelligence/platform/controller/DataStreamModelController.java
@@ -61,7 +61,7 @@ public class DataStreamModelController {
 	    	return "test";
 	    }
 	 @RequestMapping(value = "/SimPage",method = RequestMethod.GET)
-	 	public JSONObject pageable(Integer productId,Integer page,Integer number) {
+	 	public JSONObject pageable(Long productId,Integer page,Integer number) {
 		 Page<DatastreamModel> DatastreamModelPage = dsmService.queryByProductId(productId,page-1,number);
          List<DatastreamModel> DatastreamModelList = DatastreamModelPage.getContent();
          List<DataStreamModel> Data_stream_modelList = new ArrayList<>();
@@ -82,7 +82,7 @@ public class DataStreamModelController {
          return RESCODE.SUCCESS.getJSONRES(Data_stream_modelList, DatastreamModelPage.getTotalPages(), DatastreamModelPage.getTotalElements());
 	 }
 	 @RequestMapping(value ="/find_by_name",method = RequestMethod.GET)
-	 public JSONObject findByName(Integer page,Integer number,String dsmName,Integer productId) {
+	 public JSONObject findByName(Integer page,Integer number,String dsmName,Long productId) {
 		 Page<DatastreamModel> DatastreamModelPage = dsmService.queryByProductId(productId,page-1,number);
 		 return RESCODE.SUCCESS.getJSONRES(DatastreamModelPage.getContent(),DatastreamModelPage.getTotalPages(),DatastreamModelPage.getTotalElements());
 	 }
